finish up basic network which has implemented evaluation

+++ BackPropagation/ActivationFunction/Sigmoid.cs
@@ -2,6 +2,6 @@ namespace BackPropagation.ActivationFunction;
 
 public class Sigmoid : IActivationFunction
 {
-    public Func<double, double> Function { get; set; } 
-        = (value) => (1 / (1 + Math.Exp(-value)));
+    public Func<double, double> Function { get; set; } =
+        value => 1.0f / (1.0f + (float)Math.Exp(-value));
 }
